yolo-service/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
pillow>=10.0,<11.0
ultralytics>=8.3.0
